refactor: create request class for APIs
